src/main/java/faang/school/projectservice/service/stage_invitation/filter/InvitationInvitedIdFilter.java
package faang.school.projectservice.service.stage_invitation.filter;

import faang.school.projectservice.dto.invitation.StageInvitationFilterDto;
import faang.school.projectservice.model.stage_invitation.StageInvitation;

import java.util.stream.Stream;

public class InvitationInvitedIdFilter implements StageInvitationFilter {

    @Override
    public boolean isApplicable(StageInvitationFilterDto filter) {
        return filter.getInvitedIdPattern() != null;
    }

    @Override
    public Stream<StageInvitation> apply(Stream<StageInvitation> invitations, StageInvitationFilterDto filter) {
        return invitations.filter(invitation -> invitation.getInvited().getUserId().equals(filter.getInvitedIdPattern()));
    }
}
